+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Dependabot.
 - GitHub pages.
 - Discord support server. ([Invite link](https://discord.gg/SRdmrPpf2z))
+- `suggest` command.
 #### Changed
 - Embeds have now fixed colors, based on message type.